Fix getRating function and implement rating display

+++ app/src/main/java/com/finnkinoinfo/finnkinoinfo/imdbApi/ratings/ImdbRatingData.java
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ImdbRatingData {
-    public int imDb;
+    public float imDb;
 }
